+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.0.2] - 2022-10-21
+## [0.0.2] - 2022-10-22
 
 Containerized the architecture for a purely local configuration.
 
@@ -13,6 +13,8 @@ Containerized the architecture for a purely local configuration.
 
 - launch_infra.sh: a bash script to initialize infrastructural containers
 - launch_fleet.sh: a bash script to initialize Vehicle container(s)
+- teardown_infra.sh: a bash script to tear down infrastructural containers
+- teardown_fleet.sh: a bash script to tear down Vehicle container(s)
 - teardown.sh: a bash script to stop and remove infrastructural containers
 - config.master: a top-level config file to set all manner of config (JSON)
 - src/